peer: removed spring data rest

+++ peer-service/src/main/java/com/acowg/peer/entities/HasMediaOfferingFields.java
@@ -4,7 +4,7 @@ package com.acowg.peer.entities;
 public interface HasMediaOfferingFields {
     String getId();
 
-    int getSizeInBytes();
+    long getSizeInBytes();
 
     String getRelativeFilePath();
 }
